feat: removing DecoratedPressAVPlayerViewController from project and adding on Player
